Python project/requirements.txt
numpy>=1.14.5
PyQt5==5.10
pyserial==3.4
matplotlib>=3.0.3
sympy>=1.3
#virtualenv #Optional
